LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-root-dirty-expected.txt


FAIL Unrelated style mutation does not affect anonymous timeline (root) assert_equals: expected "100" but got "auto"
